+++ Lab3/main.c
@@ -15,7 +15,7 @@ void main(void* dtb)
     uart_puts("Booted! Current EL: ");
     uart_send('0' + el);
     uart_puts("\n");
-    fdt_tranverse(dtb, "linux,initrd-start", initramfs_callback);
+    //fdt_tranverse(dtb, "linux,initrd-start", initramfs_callback);
     //core_timer_enable();
     int idx = 0;
     char in_char;
