+++ Lab5/syscall.c
@@ -62,9 +62,9 @@ void exec(trapframe * sp){
     }
 
     // current is the file address
-    asm volatile ("mov x0, 0x0"); //need to check
+    asm volatile ("mov x0, 0x340"); //switch to el0 with interrupt enabled
     asm volatile ("msr spsr_el1, x0");
-    asm volatile ("msr elr_el1, %0": :"r" (new_program_pos));
+    asm volatile ("msr elr_el1, %0": :"r" (new_program_pos)); // eret address
     asm volatile ("mov x0, %0": : "r"(get_current() -> sp_el0));// user space stack
     asm volatile ("msr sp_el0, x0");
     asm volatile ("eret");
@@ -120,25 +120,27 @@ void fork(trapframe *sp){
     for(int i = 0; i< sizeof(thread); i++){
         ((char*)t)[i] = 0; 
     }
+
     t -> pid = pid;
     t -> parent = get_current() -> pid;
     t -> state = get_current() -> state;
     t -> priority = get_current() -> priority;
-    t -> sp_el1 = ((unsigned long)allocate_page(4096)) + 4096;//((unsigned long)t + 4096); //sp start from bottom
+    t -> sp_el1 = ((unsigned long)allocate_page(4096)) + 4096;// sp start from bottom
     t -> sp_el0 = ((unsigned long)allocate_page(4096)) + 4096;
     // t -> funct = fork_return;
 
-    // copy register x19~28
+    // copy callee registers
     for(int i = 0; i< sizeof(struct registers); i++){
         ((char*)(&(t -> regs)))[i] = ((char*)(&(get_current() -> regs)))[i]; 
     }
     
-    t -> regs.lr = fork_return; //need to load all and eret for child
+    t -> regs.lr = fork_return; //need to load all and eret for child trapframe
 
     //update user and kernel stack pointer
     unsigned long sp_offset = (char *)(get_current() -> sp_el1) - (char*)sp; //bottom of stack minus top of trapframe(top of stack)
     unsigned long sp_el0_offset = (char *)get_current()->sp_el0 - (char *)sp->sp_el0; //bottom of stack minus top of user stack saved by trapframe
 
+    //copy both stacks
     for(int i = 1; i<= sp_offset; i++){
         *((char *)(t -> sp_el1 - i)) = *((char *)(get_current() -> sp_el1 - i));
     }
@@ -149,9 +151,9 @@ void fork(trapframe *sp){
 
     // modify sp for load all
     t -> regs.sp = t -> sp_el1 - sp_offset; //set to the same position: top of the trapframe
-    ((trapframe*)(t -> regs.sp)) -> sp_el0 = t -> sp_el0 - sp_el0_offset; //user stack point to the same position
+    ((trapframe*)(t -> regs.sp)) -> sp_el0 = t -> sp_el0 - sp_el0_offset; //user stack point to the same position (copied trapframe)
 
-    ((trapframe*)(t -> regs.sp)) -> x[0] = 0; //return 0 for child
+    ((trapframe*)(t -> regs.sp)) -> x[0] = 0; //return 0 for child trapframe
     thread_pool[pid] = t;
     update_min_priority();
     sp -> x[0] = pid;
@@ -195,16 +197,16 @@ void sys_call(trapframe * sp){
 }
 
 void sync_exception_entry(unsigned long esr_el1, unsigned long elr_el1, trapframe* sp){
-    int ec = (esr_el1 >> 26) & 0b111111; //see 31:26 (esr_el1 spec)
     //sp is the place of "save_all"
     //x0, x1 here is the new ones for function input
+    int ec = (esr_el1 >> 26) & 0b111111; //see 31:26 (esr_el1 spec)
 
     if (ec == 0b010101) { //system call see lab3
         sys_call(sp);
     }
     else{
         /*
-        just for debug (old exception handler)
+        just for debug (lab3 exception handler)
         0b100101
         Data Abort taken without a change in Exception level.
         Used for MMU faults generated by data accesses, alignment faults
